pyproject.toml
[project]
name = "chatbot-mvp"
version = "0.1.0"
description = "智能文档问答聊天机器人 MVP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.7",
    "langchain-openai>=0.2.8",
    "langchain-community>=0.3.5",
    "langgraph>=0.2.45",
    "langsmith>=0.1.143",
    "pypdf>=4.3.1",
    "python-docx>=1.1.2",
    "markdown>=3.7",
    "chromadb>=0.5.20",
    "sentence-transformers>=3.3.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "rich>=13.9.4",
    "typer>=0.15.1",
]

[project.scripts]
chatbot = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
]
